requirements.txt
asgiref==3.7.2
Babel==2.12.1
boto3==1.26.165
botocore==1.29.165
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==41.0.1
defusedxml==0.7.1
Django==4.2.2
django-allauth==0.54.0
django-autoslug==1.9.9
django-cors-headers==4.2.0
django-countries==7.5.1
django-money==3.1.0
django-storages==1.13.2
djangorestframework==3.14.0
idna==3.4
jmespath==1.0.1
oauthlib==3.2.2
Pillow==10.0.0
py-moneyed==2.0
pycparser==2.21
PyJWT==2.7.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2023.3
requests==2.31.0
requests-oauthlib==1.3.1
s3transfer==0.6.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.0
urllib3==1.26.16
